refactor: refactored 'Work-Filters' logic
