+++ src/js/modules/work-filters/index.ts
@@ -1,0 +1,3 @@
+import WorkFilters from './ui/WorkFilters'
+
+export default WorkFilters
