+++ src/js/modules/work-filters/ui/WorkFilters.ts
@@ -1,13 +1,25 @@
-import { debounce } from '../utils/decorators'
+import { debounce } from '../../../utils/decorators'
 
+/**
+ * @class
+ * @module WorkFilters
+*/
 export default class WorkFilters {
   private _parent: HTMLElement
   private _cards: NodeListOf<HTMLElement>
 
+  /**
+   * @constructor
+   * @param {HTMLElement} workFiltersParent
+  */
   constructor(workFiltersParent: HTMLElement) {
     this._parent = workFiltersParent
   }
 
+  /**
+   * @description Initiates the 'WorkFilters' logic.
+   * @returns {void}
+  */
   public init(): void {
     this._parent.addEventListener('change', this._onChangeFilterDebounced)
     this._parent.addEventListener('reset', this._onResetFilters)
@@ -15,6 +27,10 @@ export default class WorkFilters {
     this._cards = document.querySelectorAll('[data-work-filters="work-card"]')
   }
 
+  /**
+   * Private method: [_resetCards()].
+   * @returns {void}
+  */
   private _resetCards(): void {
     this._cards.forEach((card: HTMLElement) => {
       const cardParentLi = card.parentElement as HTMLLIElement
@@ -24,16 +40,32 @@ export default class WorkFilters {
     })
   }
 
+  /**
+   * Private method: [_getCardTags()].
+   * @param {HTMLElement} card
+   * @returns {string[]}
+  */
   private _getCardTags(card: HTMLElement): string[] {
     const tags: NodeListOf<HTMLElement> = card.querySelectorAll('[data-work-filters="work-card-tag"]')
 
     return Array.from(tags).map((tag: HTMLElement) => tag.textContent!)
   }
 
+  /**
+   * Private method: [_hasWorkFilterTag()].
+   * @param {string[]} tags
+   * @param {string} workFilter
+   * @returns {boolean}
+  */
   private _hasWorkFilterTag(tags: string[], workFilter: string): boolean {
     return tags.includes(workFilter)
   }
 
+  /**
+   * Private method: [_filterCards()].
+   * @param {string} workFilter
+   * @returns {void}
+  */
   private _filterCards(workFilter: string): void {
     this._resetCards()
 
@@ -50,6 +82,13 @@ export default class WorkFilters {
     }
   }
 
+  /**
+   * @callback
+   * Private callback: [_onChangeFilter]
+   *
+   * @param {Event} e
+   * @returns {void}
+  */
   private _onChangeFilter = (e: Event): void => {
     const currTrigger: HTMLLabelElement = (e.target as HTMLElement).closest('[data-work-filters="trigger"]')!
     const currFilterId: string = currTrigger.dataset.workFilterId!
@@ -57,8 +96,20 @@ export default class WorkFilters {
     this._filterCards(currFilterId)
   }
 
+  /**
+   * @callback
+   * Private callback: [_onChangeFilterDebounced]
+   *
+   * @description Debounced original '_onChangeFilter' callback
+  */
   private _onChangeFilterDebounced = debounce(this._onChangeFilter, 250)
 
+  /**
+   * @callback
+   * Private callback: [_onResetFilters]
+   *
+   * @returns {void}
+  */
   private _onResetFilters = (): void => {
     this._resetCards()
   }
